+++ tests/e2e/game/full-game.test.js
@@ -253,6 +253,7 @@ describe("B - Full game of 18 players with all roles", () => {
                 expect(game.history[0].play.targets[0].player._id).to.equals(game.players[7]._id);
                 expect(game.history[0].play.source.name).to.equal("all");
                 expect(game.history[0].play.source.players).to.be.an("array").to.have.lengthOf(players.length);
+                expect(game.history[0].dead).to.not.exist;
                 done();
             });
     });
@@ -1305,6 +1306,10 @@ describe("B - Full game of 18 players with all roles", () => {
         expect(game.players[14].role.isRevealed).to.equals(true);
         expect(game.players[14].murdered.of).to.equals("eat");
         expect(game.players[14].murdered.by).to.equals("big-bad-wolf");
+        expect(game.history[0].dead).to.be.an("array").to.be.lengthOf(1);
+        expect(game.history[0].dead[0]._id).to.equals(players[14]._id);
+        expect(game.history[0].dead[0].murdered.of).to.equals("eat");
+        expect(game.history[0].dead[0].murdered.by).to.equals("big-bad-wolf");
         done();
     });
     it("🎲 Game is waiting for 'all' to 'vote'", done => {
@@ -1446,6 +1451,7 @@ describe("B - Full game of 18 players with all roles", () => {
                 expect(game.players[5].isAlive).to.equals(true);
                 expect(game.players[6].isAlive).to.equals(true);
                 expect(game.history[0].play.votes).to.exist;
+                expect(game.history[0].dead).to.not.exist;
                 done();
             });
     });
@@ -1550,6 +1556,9 @@ describe("B - Full game of 18 players with all roles", () => {
                 expect(game.players[6].isAlive).to.equals(false);
                 expect(game.players[6].murdered).to.deep.equals({ by: "sheriff", of: "settle-votes" });
                 expect(game.history[0].play.targets).to.exist;
+                expect(game.history[0].dead).to.be.an("array").to.be.lengthOf(1);
+                expect(game.history[0].dead[0]._id).to.equals(players[6]._id);
+                expect(game.history[0].dead[0].murdered).to.deep.equals({ by: "sheriff", of: "settle-votes" });
                 done();
             });
     });
@@ -1776,6 +1785,7 @@ describe("B - Full game of 18 players with all roles", () => {
         expect(game.players[2].isAlive).to.equals(true);
         expect(game.players[11].isAlive).to.equals(false);
         expect(game.players[11].murdered).to.deep.equals({ by: "big-bad-wolf", of: "eat" });
+        expect(game.history[0].dead).to.be.an("array").to.be.lengthOf(2);
         done();
     });
     it("🎲 Game is waiting for 'all' to 'vote'", done => {
@@ -1949,11 +1959,11 @@ describe("B - Full game of 18 players with all roles", () => {
     it("☀️ Sun is rising, little girl is eaten even if protected by guard and cupid dies from broken heart 💔", done => {
         expect(game.phase).to.equals("day");
         expect(game.players[4].attributes).to.deep.include({ attribute: "raven-marked", source: "raven", remainingPhases: 1 });
-        expect(game.players[7].attributes).to.not.deep.include({ attribute: "eaten", source: "werewolves", remainingPhases: 1 });
         expect(game.players[7].isAlive).to.equals(false);
         expect(game.players[7].murdered).to.deep.equals({ by: "werewolves", of: "eat" });
         expect(game.players[9].isAlive).to.equals(false);
         expect(game.players[9].murdered).to.deep.equals({ by: "cupid", of: "charm" });
+        expect(game.players[7].attributes).to.not.deep.include({ attribute: "eaten", source: "werewolves", remainingPhases: 1 });
         done();
     });
     it("🎲 Game is waiting for 'sheriff' to 'delegate'", done => {